Rename examples -> Ricardo
