+++ tests/test_graph/test_special.py
@@ -1,4 +1,4 @@
-from causalprog.graph.special import example_model
+from causalprog.graph.ricardo import example_model
 
 
 def test_example_model():
